requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
python-dotenv
pymysql
# Add any other packages your project needs